refactor: coordinator refactoring phase 1 - establish delegation patterns
- Extract notification methods to coordinator_callbacks.go (~170 lines)
- Extract adapter types to coordinator_adapters.go (~75 lines)
- Remove 6 deprecated functions that had fallback implementations
- Remove fallback paths - code now requires phase orchestrators
- Fix restartTask() to delegate to ExecutionOrchestrator.StartSingleTask()
- Add StartSingleTask() and RestartLoop() to ExecutionOrchestrator
- Create placeholder packages for future extraction targets

This is Phase 1 of the coordinator refactoring plan. Phases 2-5 will
migrate the remaining logic (~2,500 lines) to phase orchestrators.
