+++ internal/orchestrator/group/consolidator.go
@@ -1,0 +1,25 @@
+package group
+
+// This file is a placeholder for future group consolidation logic.
+//
+// The group consolidation logic is intended to contain:
+//   - consolidateGroupWithVerification - merge parallel task branches
+//   - getBaseBranchForGroup - determine base branch for a group
+//   - gatherTaskCompletionContextForGroup - collect context from completed tasks
+//   - getTaskBranchesForGroup - get branches for tasks in a group
+//   - buildGroupConsolidatorPrompt - build Claude prompt for consolidation
+//   - startGroupConsolidatorSession - start the consolidator instance
+//   - monitorGroupConsolidator - monitor consolidation progress
+//
+// Currently, the group consolidation logic remains in the Coordinator
+// (internal/orchestrator/coordinator.go) due to tight coupling with:
+//   - Session state (UltraPlanSession)
+//   - Orchestrator's worktree manager (wt)
+//   - Orchestrator's instance management
+//   - Base session for branch/instance operations
+//
+// Future refactoring could extract this logic here by defining interfaces
+// for the required dependencies, similar to how Tracker uses SessionData.
+//
+// The Tracker in this package already provides execution group tracking
+// that the consolidation logic could leverage.
